tests/fixtures/generators/image-curator/pyproject.toml
[project]
name = "test-image-curator"
version = "1.0.0"
description = "Intelligent test image curation for face recognition testing"
authors = [
    {name = "Chris", email = "chris@waitingforthefuture.org"}
]
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "face-recognition>=1.3.0",
    "imagehash>=4.3.1",
    "opencv-python>=4.8.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
